# app/swagger.yaml
swagger: "2.0"
info:
  version: v1.0.0
  title: "Terminal 101: Flight"
  description: Terminal 101 - Flight service
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
host: test-terminal101-flight.herokuapp.com
...
basePath: /
...
paths:
...
  /flight/search:
    # NOTE: Search a fligh
    get:
      tags:
        - Flight
      description: Search a flight
      parameters:
        - name: origin
          in: query
          type: string
          required: true
          value: MUC
          description: Flight origin code.
        - name: destination
          in: query
          type: string
          required: true
          value: IST
          description: Flight destination code.
        - name: departureDate
          in: query
          type: string
          required: true
          value: "2022-01-20"
          description: Flight departure date.
        - name: returnDate
          in: query
          type: string
          required: false
          value: "2022-02-5"
          description: Flight return date.
        - name: adults
          in: query
          type: number
          required: false
          value: 1
          description: Travelers adults count (Default equals 1).
        - name: children
          in: query
          type: number
          required: false
          description: Travelers children count (Default equals 0).
        - name: infants
          in: query
          type: number
          required: false
          description: Travelers infants count (Default equals 0).
        - name: travelClass
          in: query
          type: string
          enum: [ECONOMY, PREMIUM_ECONOMY, BUSINESS, FIRST]
          required: false
          description: Default is ECONOMY class.
        - name: segments
          in: query
          type: array
          required: false
          value:
            - DUB:NYC:2022-02-01
            - BKK:IST:2022-02-03
          items:
            type: string
          description: Keyword for search.
      responses:
        "200":
          description: OK
          schema:
            type: array
            items:
              $ref: "#/definitions/Flight"
  /flight/search/{searchId}/limits:
    # NOTE: Get popular flights
    get:
      tags:
        - Flight
      parameters:
        - name: searchId
          in: path
          required: true
          type: string
      description: Get filter limits for a search.
      responses:
        "200":
          description: OK
          schema:
            type: array
            items:
              $ref: "#/definitions/FilterLimit"
  /flight/search/{searchId}:
    # NOTE: Get popular flights
    get:
      tags:
        - Flight
      parameters:
        - name: searchId
          in: path
          required: true
          type: string
        - name: stops
          in: query
          required: false
          type: array
          items:
            type: number
        - name: airports
          in: query
          required: false
          type: array
          items:
            type: string
        - name: airlines
          in: query
          required: false
          type: array
          items:
            type: string
        - name: priceFrom
          in: query
          required: false
          type: number
        - name: priceTo
          in: query
          required: false
          type: number
        - name: departureTimeFrom
          in: query
          required: false
          type: number
        - name: departureTimeTo
          in: query
          required: false
          type: number
        - name: arrivalTimeFrom
          in: query
          required: false
          type: number
        - name: arrivalTimeTo
          in: query
          required: false
          type: number
        - name: durationFrom
          in: query
          required: false
          type: number
        - name: durationTo
          in: query
          required: false
          type: number
      description: Get search by code and filter that.
      responses:
        "200":
          description: OK
          schema:
            type: array
            items:
              $ref: "#/definitions/Flight"
  /flight/search/{searchId}/details/{flightCode}:
    # NOTE: Get popular flights
    get:
      tags:
        - Flight
      parameters:
        - name: searchId
          in: path
          required: true
          type: string
        - name: flightCode
          in: path
          required: true
          type: number
      description: Get specific flight.
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/Flight"
  /flight/search/{searchId}/details/{flightCode}/price:
    # NOTE: Get popular flights
    get:
      tags:
        - Flight
      parameters:
        - name: searchId
          in: path
          required: true
          type: string
        - name: flightCode
          in: path
          required: true
          type: number
      description: Get specific flight.
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/Flight"
...
definitions:
  Flight:
    required:
      - origin
      - destination
      - time
    properties:
      origin:
        type: string
        example: MUC
      destination:
        type: string
        example: NYC
      time:
        type: string
        example: "2022-01-13T12:15:00.000Z"
  FilterLimit:
    properties:
      stops:
        type: array
        items:
          type: number
      aircrafts:
        type: array
        items:
          type: string
      airports:
        type: array
        items:
          $ref: "#/definitions/Waypoint"
      airlines:
        type: array
        items:
          $ref: "#/definitions/Waypoint"
      price:
        $ref: "#/definitions/NumberRange"
      departureTime:
        $ref: "#/definitions/NumberRange"
      arrivalTime:
        $ref: "#/definitions/NumberRange"
      duration:
        $ref: "#/definitions/NumberRange"
...
  Waypoint:
    required:
      - code
      - name
      - description
    properties:
      code:
        type: string
        example: MUC
      name:
        type: string
        example: Munchen
      description:
        type: string
        example: Munchen airport
...
  NumberRange:
    properties:
      min:
        type: number
        example: 0
      max:
        type: number
        example: 1000
...
schemes:
  # - http
  - https
consumes:
  - application/json
produces:
  - application/json